+++ react-app/src/index.js
@@ -1,24 +1,17 @@
 import React from "react";
-// import {createRoot} from "react-dom/client";
+import {createRoot} from "react-dom/client";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
+import admin from "./admin"
 import reportWebVitals from "./reportWebVitals";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
     <React.StrictMode>
-      <Router>
-        <Routes>
-          <Route path="/" element={ <App /> }>
-          </Route>
-        </Routes>
-      </Router>
-    </React.StrictMode>,
-    document.getElementById('root')
-  );
+    <App />
+    </React.StrictMode>
+);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
